% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce_winner.R
\name{announce_winner}
\alias{announce_winner}
\title{Prepare tweet text for the winner}
\usage{
announce_winner(winner, book, book_url)
}
\arguments{
\item{winner}{as given by twitterbookdraw::draw_winner()}

\item{book}{Book title}

\item{book_url}{Book url}
}
\description{
Prepare tweet text for the winner
}
\examples{
winner <- twitterbookdraw::draw_winner()
book <- 'An Introduction to Statistical Learning: with Applications in R'
book_url <- "http://geni.us/introtostatslearning"
twitterbookdraw::announce_winner(winner, book, book_url)
}
