% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_winner.R
\name{draw_winner}
\alias{draw_winner}
\title{Draw a random follower}
\usage{
draw_winner(seed = as.numeric(Sys.Date()))
}
\arguments{
\item{seed}{Random seed (numeric)}
}
\value{
Info about the winner as a dats.frame.
}
\description{
Draw a random follower
}
