% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_august_winner.R
\name{show_august_winner}
\alias{show_august_winner}
\title{Show August winner with a gif}
\usage{
show_august_winner(winner, path = "destruction.gif")
}
\arguments{
\item{winner}{as given by twitterbookdraw::draw_winner()}

\item{path}{where to save the gif?}
}
\description{
Show August winner with a gif
}
