% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_july_winner.R
\name{show_july_winner}
\alias{show_july_winner}
\title{Show July winner with a gif}
\usage{
show_july_winner(winner, path = "july.gif")
}
\arguments{
\item{winner}{as given by twitterbookdraw::draw_winner()}

\item{path}{where to save the gif?}
}
\description{
Show July winner with a gif
}
