% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_june_winner.R
\name{show_june_winner}
\alias{show_june_winner}
\title{Create a gif for the June winner}
\usage{
show_june_winner(winner, path = "june.gif")
}
\arguments{
\item{winner}{as given by twitterbookdraw::draw_winner()}

\item{path}{where to save the gif?}
}
\description{
Create a gif for the June winner
}
